+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `Innmind\Immutable\Predicate\OrPredicate::and()`
 - `Innmind\Immutable\Predicate\AndPredicate::or()`
 - `Innmind\Immutable\Predicate\AndPredicate::and()`
+- `Innmind\Immutable\Monoid\ArrayMerge`
 
 ## 5.15.0 - 2025-05-17